Migrate entry <> namespaced range3d
